feat: serve this window's scroll group reads from a cache of main's state (PT-4275)
The renderer's scroll group module becomes what it now is: a local
representation of a service hosted elsewhere. It seeds a copy of the host's
state at startup, keeps it current from the host's events, serves the `*Sync`
readers from it, and — because the UI reads and writes a reference during
render and during a keystroke handler — predicts the host's answer to a write,
returning it immediately and reconciling afterwards. A prediction the host
declines (or a write that never reaches it) resyncs the group from the host.

Versification conversion stays here: it is a pure function of the reference and
the two projects, and its hot-path consumer is in this process, so converting
here costs one command call rather than a hop through main, and the results can
be cached against the versification subscriptions that invalidate them.

The renderer election goes with it. Main claims the name before any window
exists, so no window races for it, none can lose it when another closes, and
nothing has to watch for a handover — including the shared service's re-arm and
the e2e assertions about a window stepping aside.

`readDirection` stays in the renderer (the host exposes logical back/forward
only), and the hidden-tab case is stated where the events are consumed.

Session-URL: https://claude.ai/code/session_01Lf3XmoA9StqLNDiMBqZVTy
