+++ src/renderer/components/platform-bible-toolbar.test.tsx
@@ -96,7 +96,7 @@ vi.mock('@renderer/services/theme.service-host', () => ({
   },
 }));
 
-vi.mock('@renderer/services/scroll-group.service-host', () => ({
+vi.mock('@renderer/services/scroll-group.service', () => ({
   availableScrollGroupIds: [1, 2, 3, 4, 5],
   getReferenceHistorySync: vi.fn(() => ({ current: undefined, back: [], forward: [] })),
   navigateReferenceHistorySync: vi.fn(() => false),
