+++ src/shared/data/network-event-names.ts
@@ -9,12 +9,16 @@ const PUBLIC_MULTI_SOURCE_EVENT_NAMES = [
   'object:onDidCreateNetworkObject',
   'object:onDidDisposeNetworkObject',
   'platform.onDidChangeProjects',
-  // Scroll group and web view events are emitted by the renderers, and there is one renderer per
-  // open window. Each window navigates its own UI and opens its own web views, so each has to be
-  // able to announce what it did — under single-source semantics only the first window to start
-  // could register an emitter, and every other window's emits would be dropped.
+  // The scroll group events have one emitter, in main — but they are published in the
+  // `MultiSourceNetworkEvents` type, and multi-source is a superset of single-source semantics (a
+  // lone emitter behaves identically), so narrowing them would be a breaking change to a public type
+  // for no runtime difference.
   'scrollGroup:onDidUpdateScrRef',
   'scrollGroup:onDidChangeReferenceHistory',
+  // Web view events are emitted by the renderers, and there is one renderer per open window. Each
+  // window opens its own web views, so each has to be able to announce what it did — under
+  // single-source semantics only the first window to start could register an emitter, and every
+  // other window's emits would be dropped.
   'webView:onDidAddWebView',
   'webView:onDidOpenWebView',
   'webView:onDidUpdateWebView',
@@ -30,10 +34,10 @@ const PUBLIC_MULTI_SOURCE_EVENT_NAMES = [
  */
 const INTERNAL_MULTI_SOURCE_EVENT_NAMES = [
   'shared-store:change',
-  // Emitted by whichever renderer is tracking a project whose versification changed, so it needs
-  // the same per-window multi-source treatment as the scroll group events above. Kept internal
-  // because it is a host↔hook signal rather than part of the `@papi/*` surface — see the emitter in
-  // `scroll-group.service-host.ts`.
+  // Emitted by whichever renderer is tracking a project whose versification changed, and every
+  // renderer tracks the projects its own web views convert references for, so several can emit.
+  // Kept internal because it is a signal between the scroll group service and its consumers rather
+  // than part of the `@papi/*` surface — see the emitter in `scroll-group.service.ts`.
   'scrollGroup:onDidChangeVersification',
 ];
